ImageResizer: fix crash on using UseNewSettings
- add unit test to InteropTests
