+++ .github/actions/spell-check/expect.txt
@@ -342,6 +342,7 @@ CPPARM
 cppblog
 cppm
 cpprestsdk
+cppruntime
 cppstd
 cppwinrt
 CPPx
