+++ src/common/interop/interop-tests/InteropTests.cs
@@ -22,6 +22,9 @@ namespace Microsoft.Interop.Tests
         [TestInitialize]
         public void Initialize()
         {
+            // Make sure we don't crash
+            Assert.IsTrue(CommonManaged.ShouldNewSettingsBeUsed() == CommonManaged.ShouldNewSettingsBeUsed());
+
             ClientPipe = new TwoWayPipeMessageIPCManaged(ClientSidePipe, ServerSidePipe, null);
         }
