+++ src/common/interop/keyboard_layout.cpp
@@ -2,6 +2,10 @@
 #include "keyboard_layout_impl.h"
 #include "shared_constants.h"
 
+#include <winrt/Windows.UI.Core.h>
+
+using namespace winrt;
+
 LayoutMap::LayoutMap() :
     impl(new LayoutMap::LayoutMapImpl())
 {
